+++ src/main/kotlin/md/lhv/screening/repository/ScreeningRepository.kt
@@ -1,6 +1,6 @@
-package md.lhv.verification.repository
+package md.lhv.screening.repository
 
-import md.lhv.verification.repository.db.SanctionedPerson
+import md.lhv.screening.repository.db.SanctionedPerson
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
