+++ src/main/kotlin/md/lhv/screening/repository/db/SanctionedPerson.kt
@@ -1,4 +1,4 @@
-package md.lhv.verification.repository.db
+package md.lhv.screening.repository.db
 
 import jakarta.persistence.*
