+++ src/test/kotlin/md/lhv/screening/VerificationServiceAppTests.kt
@@ -1,4 +1,4 @@
-package md.lhv.verification
+package md.lhv.screening
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
